feat(User):完善User相关功能 (#4)
* fix:添加了maven依赖，解决依赖不能正确解析
feat:添加了统一响应体和User实体类

* feat(User):完善User登录相关,完善项目结构

* feat(User):完善User登录相关,完善项目结构

+++ bigproject/src/main/java/org/bigseven/entity/User.java
@@ -11,6 +11,7 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 
 /**
+ * 
  * @TableName user
  */
 @TableName(value ="user")
@@ -20,7 +21,8 @@ import java.time.LocalDateTime;
 @Builder
 public class User {
     @TableId(type = IdType.AUTO)
-    private Long id;
+
+    private Integer id;
     private String username;
     private String password;
     private String email;
